requirements.txt
awscrt ~=0.21.1
awsiotsdk ~=1.22.0
tzlocal ~=5.2